ai-nexus-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-dotenv
python-multipart
pydantic
twilio
